--- apps/audits/__init__.py
@@ -1,7 +1,0 @@
-# apps/audits/__init__.py
-"""
-Initialize the Audits app package.
-
-This file marks the 'audits' directory as a Python package.
-It can remain empty unless package-level code is required.
-"""
