+++ apps/certification_bodies/forms.py
@@ -3,7 +3,7 @@
 from django import forms
 from django.utils import timezone
 
-from apps.audits.models import AuditResult
+from apps.certification_bodies.models import AuditResult
 
 
 class AuditDecisionForm(forms.ModelForm):
