+++ apps/consultants/models.py
@@ -37,6 +37,9 @@ class ConsultancyFirm(models.Model):
     is_active = models.BooleanField(default=True)
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
+    temporary_field = models.CharField(
+        max_length=10, blank=True, null=True
+    )  # Temporary field to force migration
 
     def __str__(self):
         return self.name
